⚡️ FilteringBox 리스트 분리
